# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: University
  description: University specification API
  version: 1.0.0
  contact:
    name: API Botalov Anton
    email: botalovanton@gmail.com
servers:
  - url: http://localhost:8080
...
paths:
  /course:
    post:
      tags:
        - course
      summary: Add course
      description: Allows you to add a new academic discipline to the database
      operationId: addCourse
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/courseDto'
        required: true
      responses:
        201:
          description: course created
        400:
          description: validation exception
          content:
            text/plain:
              schema:
                type: string
                example: 'error message'

    get:
      tags:
        - course
      summary: All courses
      description: Returns a list of all courses
      operationId: getAllCourses
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/courseDto'
  /course/{id}:
    get:
      tags:
        - course
      summary: find course
      description: get course by id
      operationId: getCourseById
      parameters:
        - name: id
          required: true
          schema:
            type: string
            example: '999'
          in: path
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/courseDto'
        404:
          description: course not found
          content:
            text/plain:
              schema:
                type: string
                example: 'course with id 999 not found'

    delete:
      tags:
        - course
      summary: delete course
      description: delete an existing course
      operationId: deleteCourse
      parameters:
        - name: id
          required: true
          schema:
            type: string
            example: '999'
          in: path
      responses:
        204:
          description: successful course deletion

    put:
      tags:
        - course
      summary: update course
      description: update an existing course
      operationId: updateCourse
      parameters:
        - name: id
          required: true
          schema:
            type: string
            example: 999
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/courseDto'
        required: true
      responses:
        204:
          description: successful operation
        400:
          description: validation exception
          content:
            text/plain:
              schema:
                type: string
                example: 'error message'
  /course/filter:
    get:
      tags:
        - course
      summary: find by filter
      description: get a list of courses by specified filters
      operationId: getFilteredCourses
      parameters:
        - name: courseId
          required: false
          schema:
            type: string
            example: '999'
          in: query
        - name: courseName
          required: false
          schema:
            type: string
            example: 'Programming'
          in: query
        - name: departmentId
          required: false
          schema:
            type: string
            example: '999'
          in: query
        - name: departmentName
          required: false
          schema:
            type: string
            example: 'Department of economics'
          in: query
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/courseDto'
  /group:
    post:
      tags:
        - group
      summary: Add group
      description: Allows you to add a new study group to the database
      operationId: addGroup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/groupDto'
        required: true
      responses:
        201:
          description: group created
        400:
          description: validation exception
          content:
            text/plain:
              schema:
                type: string
                example: 'error message'


    get:
      tags:
        - group
      summary: All groups
      description: Returns a list of all groups
      operationId: getAllGroups
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/groupDto'
  /group/{id}:
    get:
      tags:
        - group
      summary: find group
      description: get group by id
      operationId: getGroupById
      parameters:
        - name: id
          required: true
          schema:
            type: string
            example: 999
          in: path
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/groupDto'
        404:
          description: group not found
          content:
            text/plain:
              schema:
                type: string
                example: 'group with id 999 not found'

    delete:
      tags:
        - group
      summary: delete group
      description: delete an existing group
      operationId: deleteGroup
      parameters:
        - name: id
          required: true
          schema:
            type: string
            example: 999
          in: path
      responses:
        204:
          description: successful group deletion

    put:
      tags:
        - group
      summary: update group
      description: update an existing group
      operationId: updateGroup
      parameters:
        - name: id
          required: true
          schema:
            type: string
            example: 999
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/groupDto'
        required: true
      responses:
        204:
          description: successful operation
        400:
          description: validation exception
          content:
            text/plain:
              schema:
                type: string
                example: 'error message'
  /group/filter:
    get:
      tags:
        - group
      summary: find by filter
      description: get a list of groups by specified filters
      operationId: getFilteredGroups
      parameters:
        - name: groupId
          in: query
          required: false
          schema:
            type: string
            example: '999'
        - name: groupName
          in: query
          required: false
          schema:
            type: string
            example: 'QWE-365'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/groupDto'
  /department:
    post:
      tags:
        - department
      summary: Add department
      description: Allows you to add a new department to the database
      operationId: addDepartment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/departmentDto'
        required: true
      responses:
        201:
          description: department created
        400:
          description: validation exception
          content:
            text/plain:
              schema:
                type: string
                example: 'error message'

    get:
      tags:
        - department
      summary: All departments
      description: Returns a list of all departments
      operationId: getAllDepartments
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/departmentDto'
  /department/{id}:
    get:
      tags:
        - department
      summary: find department
      description: get department by id
      operationId: getDepartmentById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: '999'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/departmentDto'
        404:
          description: department not found
          content:
            text/plain:
              schema:
                type: string
                example: 'department with id 999 not found'

    delete:
      tags:
        - department
      summary: delete department
      description: delete an existing department
      operationId: deleteDepartment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: '999'
      responses:
        204:
          description: successful department deletion

    put:
      tags:
        - department
      summary: update department
      description: update an existing department
      operationId: updateDepartment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: '999'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/departmentDto'
        required: true
      responses:
        204:
          description: successful operation
        400:
          description: validation exception
          content:
            text/plain:
              schema:
                type: string
                example: 'error message'
  /department/filter:
    get:
      tags:
        - department
      summary: find by filter
      description: get a list of departments by specified filters
      operationId: getFilteredDepartments
      parameters:
        - name: departmentId
          in: query
          required: false
          schema:
            type: string
            example: '999'
        - name: departmentName
          in: query
          required: false
          schema:
            type: string
            example: 'Department of economics'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/departmentDto'
...
components:
  schemas:
    courseDto:
      type: object
      properties:
        id:
          type: integer
          minItems: 1
          uniqueItems: true
          example: 999
        name:
          type: string
          minLength: 3
          maxLength: 30
          uniqueItems: true
          example: 'Programming'
        departmentId:
          type: integer
          minItems: 1
          example: 999
        departmentName:
          type: string
          minLength: 3
          maxLength: 15
          example: 'Department of economics'
    groupDto:
      type: object
      properties:
        id:
          type: integer
          minItems: 1
          uniqueItems: true
          example: 999
        name:
          type: string
          minLength: 3
          maxLength: 10
          uniqueItems: true
          example: 'QWE-345'
    departmentDto:
      type: object
      properties:
        id:
          type: integer
          minItems: 1
          uniqueItems: true
          example: 999
        name:
          type: string
          minLength: 3
          maxLength: 15
          uniqueItems: true
          example: 'Department of economics'
...